breaking up database files to be more manageable
