+++ monopyly/db/schema.sql
@@ -11,6 +11,7 @@ DROP TABLE IF EXISTS credit_subtransactions;
 DROP TABLE IF EXISTS credit_tags;
 DROP TABLE IF EXISTS credit_tag_links;
 
+
 /* Store user information */
 CREATE TABLE users (
   id INTEGER,
@@ -18,6 +19,7 @@ CREATE TABLE users (
   password TEXT NOT NULL,
   PRIMARY KEY (id)
 );
+
 
 /* Store information about banks */
 CREATE TABLE banks (
@@ -29,6 +31,7 @@ CREATE TABLE banks (
   UNIQUE (user_id, bank_name)
 );
 
+
 /* Store bank account type information */
 CREATE TABLE bank_account_types (
   id INTEGER,
@@ -39,13 +42,7 @@ CREATE TABLE bank_account_types (
   UNIQUE (user_id, type_name),
   UNIQUE (user_id, type_abbreviation)
 );
-/* Set some default account types (user_id=0 indicates all users) */
-INSERT INTO bank_account_types
-  (user_id, type_name, type_abbreviation)
-VALUES
-  (0, 'Savings', NULL),
-  (0, 'Checking', NULL),
-  (0, 'Certificate of Deposit', 'CD');
+
 
 /* Store bank account information */
 CREATE TABLE bank_accounts (
@@ -60,6 +57,7 @@ CREATE TABLE bank_accounts (
   UNIQUE (last_four_digits, account_type_id)
 );
 
+
 /* Store bank transaction information */
 CREATE TABLE bank_transactions (
   id INTEGER,
@@ -70,6 +68,7 @@ CREATE TABLE bank_transactions (
   PRIMARY KEY (id),
   FOREIGN KEY (account_id) REFERENCES bank_accounts (id)
 );
+
 
 /* Store credit account information */
 CREATE TABLE credit_accounts (
@@ -82,6 +81,7 @@ CREATE TABLE credit_accounts (
     ON DELETE CASCADE
 );
 
+
 /* Store credit card information */
 CREATE TABLE credit_cards (
   id INTEGER,
@@ -92,6 +92,7 @@ CREATE TABLE credit_cards (
   FOREIGN KEY (account_id) REFERENCES credit_accounts (id)
     ON DELETE CASCADE
 );
+
 
 /* Store credit card statement information */
 CREATE TABLE credit_statements (
@@ -104,6 +105,7 @@ CREATE TABLE credit_statements (
     ON DELETE CASCADE
 );
 
+
 /* Store credit card transaction information */
 CREATE TABLE credit_transactions (
   id INTEGER,
@@ -115,6 +117,7 @@ CREATE TABLE credit_transactions (
     ON DELETE CASCADE
 );
 
+
 /* Store subtransaction breakdown of transaction */
 CREATE TABLE credit_subtransactions (
   id INTEGER,
@@ -125,6 +128,7 @@ CREATE TABLE credit_subtransactions (
   FOREIGN KEY (transaction_id) REFERENCES credit_transactions (id)
     ON DELETE CASCADE
 );
+
 
 /* Store credit card transaction tags */
 CREATE TABLE credit_tags (
@@ -139,6 +143,7 @@ CREATE TABLE credit_tags (
   UNIQUE (user_id, tag_name)
 );
 
+
 /* Associate credit transactions with tags in a link table */
 CREATE TABLE credit_tag_links (
   subtransaction_id INTEGER NOT NULL,
@@ -150,107 +155,3 @@ CREATE TABLE credit_tag_links (
     ON DELETE CASCADE
 );
 
-/* Associate bank transactions and credit transactions */
-/*CREATE TABLE bank_credit_links (
-/*  bank_transaction_id INTEGER NOT NULL,
-/*  credit_transaction_id INTEGER NOT NULL,
-/*  PRIMARY KEY (bank_transaction_id, credit_transaction_id),
-/*  FOREIGN KEY (bank_transaction_id) REFERENCES bank_transactions (id)
-/*    ON DELETE CASCADE,
-/*  FOREIGN KEY (credit_transaction_id) REFERENCES credit_transactions (id)
-/*    ON DELETE CASCADE,
-/*);*/
-
-/* Prepare a view giving consolidated bank account type information */
-CREATE VIEW bank_account_types_view AS
-SELECT
-  id,
-  user_id,
-  IFNULL(type_abbreviation, type_name) type_name,
-  type_name type_full_name
-FROM bank_account_types;
-
-/* Prepare a view giving consolidated credit card transaction information */
-CREATE VIEW credit_transactions_view AS
-SELECT
-  t.id,
-  statement_id,
-  transaction_date,
-  vendor,
-  SUM(subtotal) total,
-  GROUP_CONCAT(note, '; ') notes
-FROM credit_transactions AS t
-  LEFT OUTER JOIN credit_subtransactions AS s_t
-    ON s_t.transaction_id = t.id
-GROUP BY t.id;
-
-/* Prepare a view giving enhanced credit card statement information */
-CREATE VIEW credit_statements_view AS
-WITH view AS (
-  SELECT
-    *
-  FROM (
-    SELECT
-      s.id,
-      s.issue_date,
-      c.account_id,
-      t.transaction_date,
-      /* Determine the balance on an account for each statement */
-      COALESCE(
-        SUM(subtotals) OVER (PARTITION BY account_id ORDER BY issue_date),
-        0
-      ) statement_balance,
-      /* Determine the total charges on an account for each statement */
-      COALESCE(
-        SUM(charges) OVER (PARTITION BY account_id ORDER BY issue_date),
-        0
-      ) statement_charge_total,
-      /* Determine the total payments on an account for each transaction */
-      COALESCE(
-          SUM(payments)
-          OVER (PARTITION BY account_id ORDER BY transaction_date),
-          0
-      ) daily_payment_total
-      FROM (
-        SELECT
-          statement_id,
-          transaction_date,
-          SUM(subtotal) subtotals,
-          CASE WHEN SUM(subtotal) >= 0 THEN SUM(subtotal) END charges,
-          CASE WHEN SUM(subtotal) < 0 THEN SUM(subtotal) END payments
-        FROM credit_transactions
-          INNER JOIN credit_subtransactions
-            ON credit_subtransactions.transaction_id = credit_transactions.id
-        GROUP BY transaction_id
-      ) t
-        LEFT OUTER JOIN credit_statements s
-          ON s.id = t.statement_id
-        INNER JOIN credit_cards c
-          ON c.id = s.card_id
-  )
-  GROUP BY id, daily_payment_total
-  ORDER BY transaction_date
-)
-SELECT
-  s.*,
-  v1.statement_balance balance,
-  v2.payment_date
-FROM credit_statements s
-  LEFT OUTER JOIN view v1
-    ON v1.id = s.id
-  LEFT OUTER JOIN (
-    SELECT
-      v1.id,
-      v2.transaction_date payment_date
-    FROM view v1, view v2
-    WHERE
-      /* Only compare balances for a single account */
-      v1.account_id = v2.account_id
-      /* Get times where payments offset charges (with float tolerance)*/
-      AND v1.statement_charge_total + v2.daily_payment_total < 1E-6
-      /* Exclude times where the statement charges are zero (new statements)*/
-      AND ABS(v1.statement_balance) > 1E-6
-    ORDER BY v1.transaction_date
-  ) v2
-    ON v2.id = s.id
-GROUP BY s.id;
